+++ Marlin/src/feature/bedlevel/abl/abl.cpp
@@ -47,11 +47,11 @@ static void extrapolate_one_point(const uint8_t x, const uint8_t y, const int8_t
   if (DEBUGGING(LEVELING)) {
     DEBUG_ECHOPGM("Extrapolate [");
     if (x < 10) DEBUG_CHAR(' ');
-    DEBUG_ECHO((int)x);
+    DEBUG_ECHO(x);
     DEBUG_CHAR(xdir ? (xdir > 0 ? '+' : '-') : ' ');
     DEBUG_CHAR(' ');
     if (y < 10) DEBUG_CHAR(' ');
-    DEBUG_ECHO((int)y);
+    DEBUG_ECHO(y);
     DEBUG_CHAR(ydir ? (ydir > 0 ? '+' : '-') : ' ');
     DEBUG_ECHOLNPGM("]");
   }
